Separate redux connection from details page declaration.
